+++ src/settings.js
@@ -241,7 +241,6 @@
 			const forcesave = $('#onlyofficeForcesave').is(':checked')
 			const liveViewOnShare = $('#onlyofficeLiveViewOnShare').is(':checked')
 			const help = $('#onlyofficeHelp').is(':checked')
-			const toolbarNoTabs = $('#onlyofficeToolbarNoTabs').is(':checked')
 			const reviewDisplay = $("input[type='radio'][name='reviewDisplay']:checked").attr('id').replace('onlyofficeReviewDisplay_', '')
 			const theme = $("input[type='radio'][name='theme']:checked").attr('id').replace('onlyofficeTheme_', '')
 			const unknownAuthor = $('#onlyofficeUnknownAuthor').val().trim()
@@ -266,7 +265,6 @@
 					forcesave,
 					liveViewOnShare,
 					help,
-					toolbarNoTabs,
 					reviewDisplay,
 					theme,
 					unknownAuthor,